Extract `Creators` prop to make it usable in more places
